pyproject.toml
[tool.poetry]
name = "repl_python3_beta_MindlessVerifiableObservation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
weather-api = "^1.0"
discord = "^1.7.3"
replit = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
